feat: cria producer, apos atualizacao de status para concluido é gerado uma pendencia para o stock-service (async)

+++ procurement-service/src/main/java/com/invent/io/procurement_service/dto/PurchaseResponseDto.java
@@ -9,7 +9,7 @@ import com.invent.io.procurement_service.model.Purchase;
 
 public record PurchaseResponseDto(
     Long id,
-    Integer supplierId,
+    Long supplierId,
     BigDecimal totalAmount,
     PurchaseStatus status,
     LocalDateTime purchaseDate,
